+++ CHANGELOG.md
@@ -5,8 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Fixed
-- 
+## [Unreleased]
+
+### Fixed
+
+- Fix policies
+- Fix token bug
+- Fix inventoryMiddleware
+- Update service timeout
+
 ## [2.0.4] - 2021-08-26
 
 ## [2.0.3] - 2021-08-25
@@ -28,24 +35,33 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.5] - 2021-07-27
 
 ## [0.1.3] - 2021-07-27
+
 ### Changed
+
 - Setup current linter and prettier presets.
 
 ## [0.2.0] - 2020-10-29
-### Changed 
+
+### Changed
+
 - Update documentation to include Route Params.
 
 ## [0.1.1] - 2020-06-26
+
 ### Changed
+
 - Adding Github Actions and using vtex/typescript settings.
 
 ## [0.1.0] - 2020-06-08
+
 ### Changed
+
 - Update to node builder `6.x`
 
 ## [0.0.7] - 2020-03-24
 
 ## [0.0.6] - 2020-01-11
+
 - Improving the documentation and adding more examples on the handlers.
 
 ## [0.0.5] - 2019-11-18
@@ -57,9 +73,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.0.2] - 2019-04-26
 
 ### Changed
+
 - Using new IOClient
 
 ## [0.0.1] - 2019-03-29
 
 ### Added
+
 - Initial example